+++ R/plot_emission_intensity.R
@@ -4,6 +4,12 @@
 #'   * The structure must be like [sda].
 #'   * The column `sector` must have a single value (e.g. "cement").
 #'   * (Optional) If present, the column `label` is used for data labels.
+#' @param span_5yr Logical. Use `TRUE` to restrict the time span to 5 years from
+#'   the start year (the default behavior of `qplot_emission_intensity()`), or use
+#'   `FALSE` to impose no restriction.
+#' @template convert_label
+#' @templateVar fun qplot_emission_intensity
+#' @templateVar value to_title
 #'
 #' @seealso [sda].
 #'
@@ -14,12 +20,23 @@
 #' # `data` must meet documented "Requirements"
 #' data <- subset(sda, sector == "cement")
 #' plot_emission_intensity(data)
-plot_emission_intensity <- function(data) {
+#'
+#' # plot with `qplot_emission_intensity()` parameters
+#' plot_emission_intensity(data,
+#'   span_5yr = TRUE,
+#'   convert_label = to_title
+#' )
+plot_emission_intensity <- function(data,
+                                    span_5yr = FALSE,
+                                    convert_label = identity) {
   env <- list(data = substitute(data))
   check_plot_emission_intensity(data, env = env)
 
   data %>%
-    prep_emission_intensity() %>%
+    prep_emission_intensity(
+      convert_label = convert_label,
+      span_5yr = span_5yr
+    ) %>%
     plot_emission_intensity_impl()
 }