+++ R/plot_techmix.R
@@ -9,6 +9,16 @@
 #'   * (Optional) If present, the column `label` is used for data labels.
 #'   * (Optional) If present, the column `label_tech` is used for technology
 #'   labels.
+#' @param span_5yr Logical. Use `TRUE` to restrict the time span to 5 years from
+#'   the start year (the default behavior of `qplot_techmix()`), or use
+#'   `FALSE` to impose no restriction.
+#' @template convert_label
+#' @templateVar fun qplot_techmix
+#' @templateVar value format_label_techmix
+#' @param convert_tech_label A symbol. The unquoted name of a function to apply
+#'   to technology legend labels. For example, to convert labels to uppercase
+#'   use `convert_tech_label = toupper`. To get the default behavior of
+#'   `qplot_techmix()` use `convert_tech_label = spell_out_technology`.
 #'
 #' @seealso [market_share].
 #'
@@ -26,12 +36,27 @@
 #' )
 #'
 #' plot_techmix(data)
-plot_techmix <- function(data) {
+#'
+#' # plot with `qplot_techmix()` parameters
+#' plot_techmix(data,
+#'   span_5yr = TRUE,
+#'   convert_label = format_label_techmix,
+#'   convert_tech_label = spell_out_technology
+#' )
+plot_techmix <- function(data,
+                         span_5yr = FALSE,
+                         convert_label = identity,
+                         convert_tech_label = identity) {
   env <- list(data = substitute(data))
   check_plot_techmix(data, env = env)
 
   data %>%
-    prep_techmix(env = env) %>%
+    prep_techmix(
+      convert_label = convert_label,
+      span_5yr = span_5yr,
+      convert_tech_label = convert_tech_label,
+      env = env
+    ) %>%
     plot_techmix_impl()
 }