+++ tests/testthat/test-plot_trajectory.R
@@ -78,6 +78,28 @@ test_that("Is sensitive to `center_y`", {
   expect_true(abs(start_val - lower_y_limit_centered) == abs(start_val - upper_y_limit_centered))
 })
 
+test_that("is sensitive to `convert_label`", {
+  data <- example_market_share()
+
+  labels_def <- plot_trajectory(data) %>%
+    unique_plot_data("label")
+  labels_mod <- plot_trajectory(data, convert_label = toupper) %>%
+    unique_plot_data("label")
+
+  expect_false(identical(labels_def, labels_mod))
+})
+
+test_that("is sensitive to `span_5yr`", {
+  data <- example_market_share()
+  abort_if_year_range_is_5yr_already(data)
+
+  p_f <- plot_trajectory(data, span_5yr = FALSE)
+  expect_false(diff(year_range(p_f)) == 5)
+
+  p_t <- plot_trajectory(data, span_5yr = TRUE)
+  expect_true(diff(year_range(p_t)) == 5)
+})
+
 test_that("x-axis plots year-breaks as integers (i.e. round numbers, with no-decimals)", {
   data <- market_share %>%
     filter(